+++ android/src/com/mygdx/game/persistance/Session/NNSessionDAO.java
@@ -32,7 +32,7 @@ public interface NNSessionDAO {
     int getFrameCount(int id);
 
     @Query("SELECT coordinate.id, coordinate.x, coordinate.y from coordinate, frame, frame_coordinate, session_frame WHERE session_frame.session_id = :sessionId and frame.frame_count = :frameCount LIMIT 1 OFFSET :bodyPart;")
-    NNCoordinate get_coordinates(int frameCount, int bodyPart, int sessionId);  // TODO: Vec2
+    NNCoordinate get_coordinates(int frameCount, int bodyPart, long sessionId);  // TODO: Vec2
 
     @Query("DELETE FROM session_frame")
     void nukeTable();
